fix: read Candid record field-skip count at full width (u32) (#6334)
## What

`find_field` / `skip_fields` in the RTS Candid decoder read the number
of
remaining record fields through a `*mut u8` pointer, so the count was
effectively taken modulo 256. A record whose number of trailing extra
fields is
an exact multiple of 256 produced a skip count of 0 — the decoder
skipped none
of them, left the data pointer mid-record, and went on to accept a
message a
conformant decoder would reject, mis-reading following arguments from
the
extra-field payload.

This is a decoder conformance issue: extra record fields are valid under
Candid
subtyping and must be skipped regardless of how many there are.

## Fix

- **codegen (enhanced):** store the decoded field count as a `u32` (the
classical
backend already does), so both backends write the count at the same
width into
  the slot the RTS reads back.
- **rts:** read the count as `*mut u32` instead of `*mut u8`. `u32`
rather than
`usize` matches the four valid bytes each backend stores; both count
slots are
  ≥ 4-byte aligned, so the deref is sound on wasm32 and wasm64.

## Test

`test/run/candid-record-field-skip` decodes a hand-crafted record with
256
trailing extra fields. Before the fix it mis-decoded a following
argument; now it
correctly traps (`advance out of buffer`), i.e. the malformed message is
rejected. Existing Candid tests (`candid`, `idl-decoding-bug`,
`float32-candid`)
are unaffected.

+++ rts/motoko-rts/src/idl.rs
@@ -564,7 +564,7 @@ unsafe extern "C" fn find_field(
     buf: *mut Buf,
     typtbl: *mut *mut u8,
     tag: u32,
-    n: *mut u8,
+    n: *mut u32,
 ) -> bool {
     while *n > 0 {
         let last_p = (*tb).ptr;
@@ -587,7 +587,7 @@ unsafe extern "C" fn find_field(
 }
 
 #[unsafe(no_mangle)]
-unsafe extern "C" fn skip_fields(tb: *mut Buf, buf: *mut Buf, typtbl: *mut *mut u8, n: *mut u8) {
+unsafe extern "C" fn skip_fields(tb: *mut Buf, buf: *mut Buf, typtbl: *mut *mut u8, n: *mut u32) {
     while *n > 0 {
         skip_leb128(tb);
         let it = sleb128_decode(tb);
